[Test framework MVP] AdminHeadersTest
